raised lvl cap to 255
Requires patched client exe - see client folder
Requires patched dbc for pvp difficulty
Requires database additions (stats for the new levels)
Requires changes to the field sizes (db related) since numbers can be much higher now.

+++ src/server/game/AI/NpcBots/bpet_ai.cpp
@@ -684,7 +684,7 @@ void bot_pet_ai::SetPetStats(bool force)
             if (pInfo->armor > 0)
                 myarmor = pInfo->armor;
 
-            for (uint8 i = STAT_STRENGTH; i != MAX_STATS; ++i)
+            for (uint16 i = STAT_STRENGTH; i != MAX_STATS; ++i)
                 me->SetCreateStat(Stats(i), pInfo->stats[i]);
         }
         else
